backend/requirements.txt
# 初期構成
# fastapi #==0.103.2
# uvicorn #==0.23.2
# sqlalchemy #==2.0.21
# psycopg2-binary #==2.9.9 # psycopg2-binaryはpsycopg2のスタンドアロンパッケージです。本番環境では通常のpsycopg2パッケージをインストールすることが推奨
# psycopg2
# openai #==0.28.1


aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
fastapi==0.103.2
frozenlist==1.4.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
multidict==6.0.4
openai==0.28.1
pip==23.2.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
requests==2.31.0
setuptools==65.5.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
wheel==0.41.2
yarl==1.9.2
python-dotenv